pyproject.toml
[project]
name = "box_buddy"
version = "0.0.1"
description = ""
authors = [
    {name = "Ian Day", email = "ian@ianday.me"}
]
readme = "README.md"
requires-python = ">=3.12,<4"
dependencies = [
  "argon2-cffi (==23.1.0)", # https://github.com/hynek/argon2_cffi
  "celery (==5.4.0)", # pyup: < 6.0  # https://github.com/celery/celery
  "crispy-bootstrap5 (==2024.10)", # https://github.com/django-crispy-forms/crispy-bootstrap5
  "Django (>=5.1.6,<5.2)",
  "django-allauth[socialaccount,mfa,openid] (==65.5.0)", # https://github.com/pennersr/django-allauth
  "django-bootstrap5 (==25.1)", #https://github.com/zostera/django-bootstrap5/tree/main
  "django-celery-beat (==2.7.0)", # https://github.com/celery/django-celery-beat
  "django-crispy-forms (==2.3)", # https://github.com/django-crispy-forms/django-crispy-forms
  "django-csp (==3.8)", #https://github.com/mozilla/django-csp?tab=readme-ov-file
  "django-environ (>=0.12.0,<0.13)",
  "django-health-check>=3.18.3",
  "django-import-export>=4.3.7",
  "django-model-utils (==5.0.0)", # https://github.com/jazzband/django-model-utils
  "django-redis (==5.4.0)", # https://github.com/jazzband/django-redis
  "django-simple-history>=3.8.0",
  "flower>=2.0.1", # https://github.com/mher/flower
  "gunicorn (>=23.0.0,<23.1)",
  "hiredis (==3.1.0)", # https://github.com/redis/hiredis-py
  "Pillow (==11.1.0)", # https://github.com/python-pillow/Pillow
  "psycopg2-binary>=2.9.10",
  "python-json-logger>=3.3.0",
  "python-slugify (==8.0.4)", # https://github.com/un33k/python-slugify
  "redis (==5.2.1)", # https://github.com/redis/redis-py
  "Werkzeug (>=3.1,<3.2)",
  "whitenoise (>=6.9.0,<6.10)",
]

[dependency-groups]
dev = [
  "mypy>=1.2.0",
  "ruff==0.11.5",
  "django-debug-toolbar==5.1.0", # https://github.com/jazzband/django-debug-toolbar
  "django-extensions==4.1", # https://github.com/django-extensions/django-extensions #generate model.png
  "django-coverage-plugin==3.1.0", # https://github.com/nedbat/django_coverage_plugin
  "pytest-django==4.11.1", # https://github.com/pytest-dev/pytest-django
  "watchfiles==1.0.5", # https://github.com/samuelcolvin/watchfiles
  "factory-boy==3.3.2", # https://github.com/FactoryBoy/factory_boy https://medium.com/analytics-vidhya/factoryboy-usage-cd0398fd11d2
  "pre-commit>=4.2.0",
  "django-stubs[compatible-mypy]>=5.1.3",
  "types-django-import-export>=4.3.0.20250218",
  "djlint>=1.36.4"
]
docs = [
  "mkdocs  (>=1.6,<1.7)",
  "mkdocs-autorefs (>=1.2,<1.3)",
  "mkdocs-get-deps (>=0.2,<0.3)",
  "mkdocs-glightbox (>=0.4,<0.5)",
  "mkdocs-material (>=9.5,<9.6)",
  "mkdocs-material-extensions (>=1.3,<1.4)",
  "mkdocstrings[python] (>=0.26,<0.27)"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.mypy]
python_executable = "/app_env/bin/python"
plugins = [
    "mypy_django_plugin.main",
]
disable_error_code = ["import-untyped"]

[[tool.mypy.overrides]]
# Django migrations should not produce any errors:
module = "*.migrations.*"
ignore_errors = true

[tool.django-stubs]
django_settings_module = "box_buddy.settings"

[tool.coverage.run]
include = ["box_buddy/**"]
omit = ["*/migrations/*", "*/tests/*"]
plugins = ["django_coverage_plugin"]

[tool.coverage.django_coverage_plugin]
template_extensions = 'html, txt, tex, email'

[tool.ruff]
target-version = "py312"
line-length = 120
extend-exclude = [
    "*/migrations/*.py",
    "staticfiles/*",
]

[tool.ruff.lint]
select = [
  "F",
  "E",
  "W",
  "C90",
  "I",
  "N",
  "UP",
  "YTT",
  "ASYNC",
  "S",
  "BLE",
  "FBT",
  "B",
  "A",
  "COM",
  "C4",
  "DTZ",
  "T10",
  "DJ",
  "EM",
  "EXE",
  "FA",
  'ISC',
  "ICN",
  "G",
  'INP',
  'PIE',
  "T20",
  'PYI',
  'PT',
  "Q",
  "RSE",
  "RET",
  "SLF",
  "SLOT",
  "SIM",
  "TID",
  "TC",
  "INT",
  "PTH",
  "ERA",
  "PD",
  "PGH",
  "PL",
  "TRY",
  "FLY",
  "PERF",
  "RUF",
]
ignore = [
  "TRY003",
  "EM101",
  "S101", # Use of assert detected https://docs.astral.sh/ruff/rules/assert/
  "RUF012", # Mutable class attributes should be annotated with `typing.ClassVar`
  "SIM102", # sometimes it's better to nest
  "COM812",
  "UP038", # Checks for uses of isinstance/issubclass that take a tuple
          # of types for comparison.
          # Deactivated because it can make the code slow:
          # https://github.com/astral-sh/ruff/issues/7871
]
# The fixes in extend-unsafe-fixes will require
# provide the `--unsafe-fixes` flag when fixing.
extend-unsafe-fixes = [
  "UP038",
]

[tool.ruff.lint.isort]
force-single-line = true


[tool.pytest.ini_options]
minversion = "6.0"
addopts = "--ds=box_buddy.settings --reuse-db --import-mode=importlib"
python_files = [
    "tests.py",
    "test_*.py",
]
DJANGO_SETTINGS_MODULE = "box_buddy.settings"


[tool.djlint]
blank_line_after_tag = "load,extends"
close_void_tags = true
format_css = true
format_js = true
ignore = "H006,H030,H031,T002"
include = "H017,H035"
indent = 2
max_line_length = 119
profile = "django"

[tool.djlint.css]
indent_size = 2

[tool.djlint.js]
indent_size = 2
